changed User's Id to int and updated migration

+++ backend/Squirrel.Core/Squirrel.Core.DAL/Entities/User.cs
@@ -2,7 +2,7 @@
 
 namespace Squirrel.Core.DAL.Entities;
 
-public sealed class User : Entity<long>
+public sealed class User : Entity<int>
 {
     public string Username { get; set; }
     public string FirstName { get; set; }
