+++ backend/Squirrel.Core/Squirrel.Core.DAL/Migrations/20230823132632_AddedUsers.cs
@@ -12,7 +12,7 @@ namespace Squirrel.Core.DAL.Migrations
                 name: "Users",
                 columns: table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Username = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     FirstName = table.Column<string>(type: "nvarchar(max)", nullable: false),
